added a function to writeMailto with content, problem discovered: i2cwrite have limited data transfered

+++ Inc/i2c.h
@@ -11,8 +11,8 @@
 #include "stdint.h"
 //#include "stm32g0xx_hal_def.h"
 
-void I2CRead(uint16_t devAddress,uint16_t memAddress, uint8_t *data, int n);
-void I2CWrite(uint16_t devAddress,uint16_t memAddress, uint8_t *data, int n);
+void I2CRead(uint16_t devAddress,uint16_t memAddress, uint8_t *data, uint16_t n);
+void I2CWrite(uint16_t devAddress,uint16_t memAddress, uint8_t *data, uint16_t n);
 void NFC_Delay(uint32_t delay);
 
 #endif /* I2C_H_ */
