+++ Src/i2c.c
@@ -7,14 +7,13 @@
 
 #include "main.h"
 #include "i2c.h"
-//#include "stm32g0xx_hal.h"
 
-void I2CRead(uint16_t devAddress,uint16_t memAddress, uint8_t *data, int n){
+void I2CRead(uint16_t devAddress,uint16_t memAddress, uint8_t *data, uint16_t n){
 	HAL_I2C_Mem_Read(&hi2c1, devAddress, memAddress, 2, data, n, 50);
 	HAL_Delay(10);
 }
 
-void I2CWrite(uint16_t devAddress,uint16_t memAddress, uint8_t *data, int n){
+void I2CWrite(uint16_t devAddress,uint16_t memAddress, uint8_t *data, uint16_t n){
 	HAL_I2C_Mem_Write(&hi2c1, devAddress, memAddress, 2, data, n, 50);
 	HAL_Delay(10);
 }
